fix: code refacto
